+++ spec/osm/activity_spec.rb
@@ -93,4 +93,29 @@ describe "Using The API" do
     activity.valid?.should be_true
   end
 
+
+    it "Add activity to programme (succeded)" do
+      url = 'https://www.onlinescoutmanager.co.uk/programme.php?action=addActivityToProgramme'
+      post_data = {
+        'apiid' => @CONFIGURATION[:api][:osm][:id],
+        'token' => @CONFIGURATION[:api][:osm][:token],
+        'userid' => 'user_id',
+        'secret' => 'secret',
+        'meetingdate' => '2000-01-02',
+        'sectionid' => 1,
+        'activityid' => 2,
+        'notes' => 'Notes',
+      }
+
+      HTTParty.should_receive(:post).with(url, {:body => post_data}) { DummyHttpResult.new(:response=>{:code=>'200', :body=>'{"result":0}'}) }
+      activity = Osm::Activity.new(:id => 2)
+      activity.add_to_programme(@api, 1, Date.new(2000, 1, 2), 'Notes').should be_true
+    end
+
+    it "Add activity to programme (failed)" do
+      HTTParty.should_receive(:post) { DummyHttpResult.new(:response=>{:code=>'200', :body=>'{"result":1}'}) }
+      activity = Osm::Activity.new(:id => 2)
+      activity.add_to_programme(@api, 1, Date.new(2000, 1, 2), 'Notes').should be_false
+    end
+
 end