+++ spec/osm/event_spec.rb
@@ -134,7 +134,7 @@ describe "Event" do
       event.reminders.should == false
       event.columns[0].id.should == 'f_1'
       event.columns[0].name.should == 'Name'
-      event.columns[0].parent_label.should == 'Label'
+      event.columns[0].label.should == 'Label'
       event.valid?.should be_true
     end
 
@@ -434,7 +434,10 @@ describe "Event" do
       event = Osm::Event.new(:id => 2, :section_id => 1)
       event.should_not be_nil
       event.add_column(@api, 'Test name', 'Test label').should be_true
-      event.columns.should == [Osm::Event::Column.new(:id=>'f_1', :name=>'Test name', :parent_label=>'Test label')]
+      column = event.columns[0]
+      column.id.should == 'f_1'
+      column.name.should == 'Test name'
+      column.label.should == 'Test label'
     end
 
     it "Add column (failed)" do
@@ -443,6 +446,77 @@ describe "Event" do
       event = Osm::Event.new(:id => 2, :section_id => 1)
       event.should_not be_nil
       event.add_column(@api, 'Test name', 'Test label').should be_false
+    end
+
+
+    it "Update column (succeded)" do
+      url = 'https://www.onlinescoutmanager.co.uk/events.php?action=renameColumn&sectionid=1&eventid=2'
+      post_data = {
+        'apiid' => @CONFIGURATION[:api][:osm][:id],
+        'token' => @CONFIGURATION[:api][:osm][:token],
+        'userid' => 'user_id',
+        'secret' => 'secret',
+        'columnId' => 'f_1',
+        'columnName' => 'New name',
+        'pL' => 'New label',
+      }
+      body = {
+        'eventid' => '2',
+        'config' => '[{"id":"f_1","name":"New name","pL":"New label"}]'
+      }
+      HTTParty.should_receive(:post).with(url, {:body => post_data}) { DummyHttpResult.new(:response=>{:code=>'200', :body=>body.to_json}) }
+
+      event = Osm::Event.new(:id => 2, :section_id => 1)
+      event.columns = [Osm::Event::Column.new(:id => 'f_1', :event => event)]
+      column = event.columns[0]
+      column.name = 'New name'
+      column.label = 'New label'
+
+      column.update(@api).should be_true
+
+      column.name.should == 'New name'
+      column.label.should == 'New label'
+      event.columns[0].name.should == 'New name'
+      event.columns[0].label.should == 'New label'
+    end
+
+    it "Update column (failed)" do
+      HTTParty.should_receive(:post) { DummyHttpResult.new(:response=>{:code=>'200', :body=>'{"config":"[]"}'}) }
+
+      event = Osm::Event.new(:id => 2, :section_id => 1)
+      column = Osm::Event::Column.new(:id => 'f_1', :event => event)
+      event.columns = [column]
+      column.update(@api).should be_false
+    end
+
+
+    it "Delete column (succeded)" do
+      url = 'https://www.onlinescoutmanager.co.uk/events.php?action=deleteColumn&sectionid=1&eventid=2'
+      post_data = {
+        'apiid' => @CONFIGURATION[:api][:osm][:id],
+        'token' => @CONFIGURATION[:api][:osm][:token],
+        'userid' => 'user_id',
+        'secret' => 'secret',
+        'columnId' => 'f_1'
+      }
+
+      HTTParty.should_receive(:post).with(url, {:body => post_data}) { DummyHttpResult.new(:response=>{:code=>'200', :body=>'{"eventid":"2","config":"[]"}'}) }
+
+      event = Osm::Event.new(:id => 2, :section_id => 1)
+      column = Osm::Event::Column.new(:id => 'f_1', :event => event)
+      event.columns = [column]
+
+      column.delete(@api).should be_true
+      event.columns.should == []
+    end
+
+    it "Delete column (failed)" do
+      HTTParty.should_receive(:post) { DummyHttpResult.new(:response=>{:code=>'200', :body=>'{"config":"[{\"id\":\"f_1\"}]"}'}) }
+
+      event = Osm::Event.new(:id => 2, :section_id => 1)
+      column = Osm::Event::Column.new(:id => 'f_1', :event => event)
+      event.columns = [column]
+      column.delete(@api).should be_false
     end
 
   end