+++ spec/osm/flexi_record_spec.rb
@@ -85,6 +85,190 @@ describe "Flexi Record" do
       fields[3].id.should == 'f_2'
     end
 
+    it "Add field (success)" do
+      url = "https://www.onlinescoutmanager.co.uk/extras.php?action=addColumn&sectionid=1&extraid=2"
+
+      post_data = {
+        'apiid' => @CONFIGURATION[:api][:osm][:id],
+        'token' => @CONFIGURATION[:api][:osm][:token],
+        'userid' => 'user_id',
+        'secret' => 'secret',
+        'columnName' => 'name',
+      }
+
+      data = {
+        "extraid" => "2",
+        "sectionid" => "1",
+        "name" => "A Flexi Record",
+        "config" => "[{\"id\":\"f_1\",\"name\":\"name\",\"width\":\"150\"}]",
+        "total" => "none",
+        "extrafields" => "[]",
+        "structure" => [
+          {
+            "rows" => [
+              {"name" => "First name","field" => "firstname","width" => "150px"},
+              {"name" => "Last name","field" => "lastname","width" => "150px"},
+            ],
+            "noscroll" => true
+          },
+          {"rows" => [
+            {"name" => "name","field" => "f_1","width" => "150px","editable" => true},
+          ]}
+        ]
+      }
+      HTTParty.should_receive(:post).with(url, {:body => post_data}) { DummyHttpResult.new(:response=>{:code=>'200', :body=>data.to_json}) }
+
+      Osm::FlexiRecord.add_field(@api, 1, 2, 'name').should be_true
+    end
+
+    it "Add field (failure)" do
+      data = {
+        "extraid" => "2",
+        "sectionid" => "1",
+        "name" => "A Flexi Record",
+        "config" => "[]",
+        "total" => "none",
+        "extrafields" => "[]",
+        "structure" => [
+          {
+            "rows" => [
+              {"name" => "First name","field" => "firstname","width" => "150px"},
+              {"name" => "Last name","field" => "lastname","width" => "150px"},
+            ],
+            "noscroll" => true
+          },
+          {"rows" => [
+          ]}
+        ]
+      }
+      HTTParty.should_receive(:post) { DummyHttpResult.new(:response=>{:code=>'200', :body=>data.to_json}) }
+
+      Osm::FlexiRecord.add_field(@api, 1, 2, 'name').should be_false
+    end
+
+    it "Update field (success)" do
+      url = "https://www.onlinescoutmanager.co.uk/extras.php?action=renameColumn&sectionid=1&extraid=2"
+
+      post_data = {
+        'apiid' => @CONFIGURATION[:api][:osm][:id],
+        'token' => @CONFIGURATION[:api][:osm][:token],
+        'userid' => 'user_id',
+        'secret' => 'secret',
+        'columnId' => 'f_1',
+        'columnName' => 'name',
+      }
+
+      data = {
+        "extraid" => "2",
+        "sectionid" => "1",
+        "name" => "A Flexi Record",
+        "config" => "[{\"id\":\"f_1\",\"name\":\"name\",\"width\":\"150\"}]",
+        "total" => "none",
+        "extrafields" => "[]",
+        "structure" => [
+          {
+            "rows" => [
+              {"name" => "First name","field" => "firstname","width" => "150px"},
+              {"name" => "Last name","field" => "lastname","width" => "150px"},
+            ],
+            "noscroll" => true
+          },
+          {"rows" => [
+            {"name" => "name","field" => "f_1","width" => "150px","editable" => true},
+          ]}
+        ]
+      }
+      HTTParty.should_receive(:post).with(url, {:body => post_data}) { DummyHttpResult.new(:response=>{:code=>'200', :body=>data.to_json}) }
+
+      Osm::FlexiRecord.update_field(@api, 1, 2, 'f_1', 'name').should be_true
+    end
+
+    it "Update field (failure)" do
+      data = {
+        "extraid" => "2",
+        "sectionid" => "1",
+        "name" => "A Flexi Record",
+        "config" => "[]",
+        "total" => "none",
+        "extrafields" => "[]",
+        "structure" => [
+          {
+            "rows" => [
+              {"name" => "First name","field" => "firstname","width" => "150px"},
+              {"name" => "Last name","field" => "lastname","width" => "150px"},
+            ],
+            "noscroll" => true
+          },
+          {"rows" => [
+          ]}
+        ]
+      }
+      HTTParty.should_receive(:post) { DummyHttpResult.new(:response=>{:code=>'200', :body=>data.to_json}) }
+
+      Osm::FlexiRecord.update_field(@api, 1, 2, 'f_1', 'name').should be_false
+    end
+
+
+    it "Delete field (success)" do
+      url = "https://www.onlinescoutmanager.co.uk/extras.php?action=deleteColumn&sectionid=1&extraid=2"
+
+      post_data = {
+        'apiid' => @CONFIGURATION[:api][:osm][:id],
+        'token' => @CONFIGURATION[:api][:osm][:token],
+        'userid' => 'user_id',
+        'secret' => 'secret',
+        'columnId' => 'f_1',
+      }
+
+      data = {
+        "extraid" => "2",
+        "sectionid" => "1",
+        "name" => "A Flexi Record",
+        "config" => "[]",
+        "total" => "none",
+        "extrafields" => "[]",
+        "structure" => [
+          {
+            "rows" => [
+              {"name" => "First name","field" => "firstname","width" => "150px"},
+              {"name" => "Last name","field" => "lastname","width" => "150px"},
+            ],
+            "noscroll" => true
+          },
+          {"rows" => []}
+        ]
+      }
+      HTTParty.should_receive(:post).with(url, {:body => post_data}) { DummyHttpResult.new(:response=>{:code=>'200', :body=>data.to_json}) }
+
+      Osm::FlexiRecord.delete_field(@api, 1, 2, 'f_1').should be_true
+    end
+
+    it "Delete field (failure)" do
+      data = {
+        "extraid" => "2",
+        "sectionid" => "1",
+        "name" => "A Flexi Record",
+        "config" => "[{\"id\":\"f_1\",\"name\":\"name\",\"width\":\"150\"}]",
+        "total" => "none",
+        "extrafields" => "[]",
+        "structure" => [
+          {
+            "rows" => [
+              {"name" => "First name","field" => "firstname","width" => "150px"},
+              {"name" => "Last name","field" => "lastname","width" => "150px"},
+            ],
+            "noscroll" => true
+          },
+          {"rows" => [
+          ]}
+        ]
+      }
+      HTTParty.should_receive(:post) { DummyHttpResult.new(:response=>{:code=>'200', :body=>data.to_json}) }
+
+      Osm::FlexiRecord.delete_field(@api, 1, 2, 'f_1').should be_false
+    end
+
+
     it "Fetch Data" do
       data = {
         'identifier' => 'scoutid',
@@ -122,6 +306,47 @@ describe "Flexi Record" do
         'f_2' => 'B',
       }
       record.valid?.should be_true
+    end
+
+
+    it "Update data (success)" do
+      url = "https://www.onlinescoutmanager.co.uk/extras.php?action=updateScout"
+
+      post_data = {
+        'apiid' => @CONFIGURATION[:api][:osm][:id],
+        'token' => @CONFIGURATION[:api][:osm][:token],
+        'userid' => 'user_id',
+        'secret' => 'secret',
+        'termid' => 1,
+        'scoutid' => 2,
+        'column' => 'f_1',
+        'value' => 'value',
+        'sectionid' => 3,
+        'extraid' => 4,
+      }
+
+      data = {
+        'items' => [
+          {'f_1' => 'value', 'scoutid' => '2'},
+        ]
+      }
+      HTTParty.should_receive(:post).with(url, {:body => post_data}) { DummyHttpResult.new(:response=>{:code=>'200', :body=>data.to_json}) }
+      Osm::Term.stub(:get_current_term_for_section) { Osm::Term.new(:id => 1) }
+
+      Osm::FlexiRecord.update_data(@api, 3, 4, 2, 'f_1', 'value').should be_true
+    end
+
+    it "Update data (failed)" do
+      data = {
+        'items' => [
+          {'f_1' => 'old value', 'scoutid' => '2'},
+        ]
+      }
+
+      HTTParty.stub(:post) { DummyHttpResult.new(:response=>{:code=>'200', :body=>data.to_json}) }
+      Osm::Term.stub(:get_current_term_for_section) { Osm::Term.new(:id => 1) }
+
+      Osm::FlexiRecord.update_data(@api, 3, 4, 2, 'f_1', 'value').should be_false
     end